Stop autowiring the CertificateController.
Autowiring the CertificateController meant that there had to be a bean
with a particular name, which meant hard-wiring its resource type in
the configuration. Setting a property on the certificate controller
bean instead allows us to generalise the signing key and certificate
properties. In most cases, of course, they will still end up as
FileSystemResources.